Exception handler cleanup: 1. Remove superfluous EspExceptionHandler abstractions. 2. Remove deprecated FlinkEspExceptionHandler implementations.
