+++ api/src/main/scala/pl/touk/nussknacker/engine/api/exception/EspExceptionInfo.scala
@@ -1,0 +1,5 @@
+package pl.touk.nussknacker.engine.api.exception
+
+import pl.touk.nussknacker.engine.api.Context
+
+case class EspExceptionInfo[T <: Throwable](nodeId: Option[String], throwable: T, context: Context) extends Serializable
