+++ api/src/main/scala/pl/touk/nussknacker/engine/api/process/ProcessConfigCreator.scala
@@ -1,7 +1,6 @@
 package pl.touk.nussknacker.engine.api.process
 
 import pl.touk.nussknacker.engine.api._
-import pl.touk.nussknacker.engine.api.exception.ExceptionHandlerFactory
 import pl.touk.nussknacker.engine.api.signal.ProcessSignalSender
 
 /**
