+++ engine/flink/management/java_sample/src/main/java/pl/touk/nussknacker/engine/management/javasample/TestProcessConfigCreator.java
@@ -3,13 +3,11 @@ package pl.touk.nussknacker.engine.management.javasample;
 import pl.touk.nussknacker.engine.api.CustomStreamTransformer;
 import pl.touk.nussknacker.engine.api.ProcessListener;
 import pl.touk.nussknacker.engine.api.Service;
-import pl.touk.nussknacker.engine.api.exception.ExceptionHandlerFactory;
 import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies;
 import pl.touk.nussknacker.engine.api.process.SinkFactory;
 import pl.touk.nussknacker.engine.api.process.SourceFactory;
 import pl.touk.nussknacker.engine.api.process.WithCategories;
 import pl.touk.nussknacker.engine.api.signal.ProcessSignalSender;
-import pl.touk.nussknacker.engine.flink.util.exception.ConfigurableExceptionHandlerFactory;
 import pl.touk.nussknacker.engine.javaapi.process.ExpressionConfig;
 import pl.touk.nussknacker.engine.javaapi.process.ProcessConfigCreator;
 
@@ -49,11 +47,6 @@ public class TestProcessConfigCreator implements ProcessConfigCreator {
     @Override
     public Collection<ProcessListener> listeners(ProcessObjectDependencies processObjectDependencies) {
         return Collections.emptyList();
-    }
-
-    @Override
-    public ExceptionHandlerFactory exceptionHandlerFactory(ProcessObjectDependencies processObjectDependencies) {
-        return ConfigurableExceptionHandlerFactory.apply(processObjectDependencies);
     }
 
     @Override
