+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/ProcessObjectDefinitionExtractor.scala
@@ -2,7 +2,6 @@ package pl.touk.nussknacker.engine.definition
 
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.NodeId
-import pl.touk.nussknacker.engine.api.exception.{EspExceptionHandler, ExceptionHandlerFactory}
 import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.api.signal.ProcessSignalSender
 
@@ -29,7 +28,6 @@ object ProcessObjectDefinitionExtractor {
 
   val source = SourceProcessObjectDefinitionExtractor
   val sink = new ProcessObjectDefinitionExtractor[SinkFactory, Sink]
-  val exceptionHandler = new ProcessObjectDefinitionExtractor[ExceptionHandlerFactory, EspExceptionHandler]
   val customNodeExecutor: CustomStreamTransformerExtractor.type = CustomStreamTransformerExtractor
   val service: MethodDefinitionExtractor[Service] = DefaultServiceInvoker.Extractor
   val signals: SignalsDefinitionExtractor.type = SignalsDefinitionExtractor
